增加了回复功能

+++ src/main/java/com/zjl/community/controller/PublishController.java
@@ -1,7 +1,6 @@
 package com.zjl.community.controller;
 
 import com.zjl.community.dto.QuestionDTO;
-import com.zjl.community.mapper.QuestionMapper;
 import com.zjl.community.model.Question;
 import com.zjl.community.model.User;
 import com.zjl.community.service.QuestionService;
@@ -23,7 +22,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                         Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle());
@@ -44,7 +43,7 @@ public class PublishController {
             @RequestParam(value = "title",required=false) String title,
             @RequestParam(value = "description",required=false) String description,
             @RequestParam(value = "tag",required=false) String tag,
-            @RequestParam(value = "id",required = false) Integer id,
+            @RequestParam(value = "id",required = false) Long id,
             HttpServletRequest request,
             Model model){
         model.addAttribute("title",title);
